+++ .changeset/ninety-islands-move.md
@@ -1,0 +1,7 @@
+---
+'@mastra/playground-ui': patch
+'mastra': patch
+'@mastra/core': patch
+---
+
+adjust the way we display scorers in agent metadata
