+++ packages/playground-ui/src/domains/scores/index.ts
@@ -1,4 +1,3 @@
-export * from './components/scorer-list';
 export * from './components/score-dialog';
 export * from './components/scores-tools';
 export * from './hooks/use-scorers';
